+++ bubbles/api/__init__.py
@@ -12,7 +12,7 @@ from fastapi import HTTPException, Request
 from fastapi.security import OAuth2PasswordBearer
 
 from bubbles.bubbles import Bubbles
-from bubbles.backend.auth import JWT, JWTDenyList, JWTMgr
+from bubbles.auth import JWT, JWTDenyList, JWTMgr
 
 
 class JWTAuthSchema(OAuth2PasswordBearer):
