+++ bubbles/api/cluster.py
@@ -5,22 +5,21 @@
 # License as published by the Free Software Foundation; either
 # version 2.1 of the License, or (at your option) any later version.
 #
-from typing import Callable, List
+from typing import Callable
 
 from fastapi import APIRouter, Depends, Request
 
 from bubbles.bubbles import Bubbles
-from bubbles.backend.api import jwt_auth_scheme
-from bubbles.backend.models.host import HostModel
+from bubbles.api import jwt_auth_scheme
+from bubbles.models.df import ClusterUsageStatsModel
 
-router = APIRouter(prefix="/host", tags=["host"])
+router = APIRouter(prefix="/cluster", tags=["cluster"])
 
 
-@router.get("/", name="Get list of hosts", response_model=List[HostModel])
-async def get_hosts(
+@router.get("/df", response_model=ClusterUsageStatsModel)
+async def get_df(
     request: Request, _: Callable = Depends(jwt_auth_scheme)
-) -> List[HostModel]:
+) -> ClusterUsageStatsModel:
     bubbles: Bubbles = request.app.state.bubbles
-    assert bubbles.ctrls.rest_api_proxy is not None
-    response = bubbles.ctrls.rest_api_proxy.request("GET", "/api/host")
-    return [HostModel.parse_obj(host) for host in response]
+    assert bubbles.ctrls.cluster is not None
+    return bubbles.ctrls.cluster.df()
