+++ bubbles/api/host.py
@@ -5,21 +5,22 @@
 # License as published by the Free Software Foundation; either
 # version 2.1 of the License, or (at your option) any later version.
 #
-from typing import Callable
+from typing import Callable, List
 
 from fastapi import APIRouter, Depends, Request
 
 from bubbles.bubbles import Bubbles
-from bubbles.backend.api import jwt_auth_scheme
-from bubbles.backend.models.df import ClusterUsageStatsModel
+from bubbles.api import jwt_auth_scheme
+from bubbles.models.host import HostModel
 
-router = APIRouter(prefix="/cluster", tags=["cluster"])
+router = APIRouter(prefix="/host", tags=["host"])
 
 
-@router.get("/df", response_model=ClusterUsageStatsModel)
-async def get_df(
+@router.get("/", name="Get list of hosts", response_model=List[HostModel])
+async def get_hosts(
     request: Request, _: Callable = Depends(jwt_auth_scheme)
-) -> ClusterUsageStatsModel:
+) -> List[HostModel]:
     bubbles: Bubbles = request.app.state.bubbles
-    assert bubbles.ctrls.cluster is not None
-    return bubbles.ctrls.cluster.df()
+    assert bubbles.ctrls.rest_api_proxy is not None
+    response = bubbles.ctrls.rest_api_proxy.request("GET", "/api/host")
+    return [HostModel.parse_obj(host) for host in response]
