+++ bubbles/api/users.py
@@ -10,8 +10,8 @@ from typing import Callable, List
 from fastapi import APIRouter, Depends, Request
 
 from bubbles.bubbles import Bubbles
-from bubbles.backend.api import jwt_auth_scheme
-from bubbles.backend.models.user import UserModel
+from bubbles.api import jwt_auth_scheme
+from bubbles.models.user import UserModel
 
 router = APIRouter(prefix="/users", tags=["users"])
