+++ bubbles/errors.py
