+++ bubbles/models/auth.py
