+++ bubbles/typings/ceph_module.pyi
