+++ bubbles/typings/mgr_module.pyi
