+++ bubbles/utils.py
