+++ tox.ini
@@ -23,7 +23,7 @@ deps =
     {[base-test]deps}
 commands =
     pytest {posargs: \
-        backend/ }
+        tests/ }
 
 [testenv:mypy]
 basepython = python3
@@ -32,7 +32,7 @@ deps =
     {[base-test]deps}
     mypy
 setenv =
-    MYPYPATH=../:backend/typings
+    MYPYPATH=../:bubbles/typings
 commands =
     mypy --config-file=mypy.ini \
             -m bubbles
@@ -42,9 +42,7 @@ basepython = python3
 deps =
     black
 modules =
-    bubbles.py
-    module.py
-    backend
+    bubbles
 commands =
     black {posargs:{[testenv:lint]modules}}
 
@@ -54,9 +52,7 @@ skip_install = true
 deps =
     black
 modules =
-    bubbles.py
-    module.py
-    backend
+    bubbles
 commands =
     black --check --diff {posargs:{[testenv:lint]modules}}
